Bugfix: dont show the Shared Folders part if there is no existing Shared Folder...
git-svn-id: file:///home/code/svn/ovd/trunk@1435 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ SessionManager/web/admin/usersgroup.php
@@ -801,50 +801,53 @@ echo '</form>';
     }
     echo '</table>';
     echo '</div>';
-	echo '<br />';
   }
 
-    echo '<div>';
-    echo '<h2>'._('Shared folders').'</h1>';
+    $all_sharedfolders = SharedFolders::getAll();
 
-	$all_sharedfolders = SharedFolders::getAll();
+    if (count($all_sharedfolders) > 0) {
+		$available_sharedfolders = array();
+		$used_sharedfolders = Abstract_SharedFolder::load_by_usergroup_id($group->getUniqueID());
+		foreach ($all_sharedfolders as $sharedfolder) {
+			if (in_array($sharedfolder->id, array_keys($used_sharedfolders)))
+				continue;
 
-	$available_sharedfolders = array();
-	$used_sharedfolders = Abstract_SharedFolder::load_by_usergroup_id($group->getUniqueID());
-	foreach ($all_sharedfolders as $sharedfolder) {
-		if (in_array($sharedfolder->id, array_keys($used_sharedfolders)))
-			continue;
+			$available_sharedfolders[] = $sharedfolder;
+		}
 
-		$available_sharedfolders[] = $sharedfolder;
-	}
+        echo '<br />';
+		echo '<div>';
+		echo '<h2>'._('Shared folders').'</h1>';
 
-	echo '<table border="0" cellspacing="1" cellpadding="3">';
-	foreach ($used_sharedfolders as $sharedfolder) {
-		echo '<tr>';
-		echo '<td><a href="sharedfolders.php?action=manage&amp;id='.$sharedfolder->id.'">'.$sharedfolder->name.'</a></td>';
-		echo '<td><form action="actions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to delete this shared folder access?').'\');">';
-		echo '<input type="hidden" name="name" value="SharedFolder_ACL" />';
-		echo '<input type="hidden" name="action" value="del" />';
-		echo '<input type="hidden" name="sharedfolder_id" value="'.$sharedfolder->id.'" />';
-		echo '<input type="hidden" name="usergroup_id" value="'.$group->getUniqueID().'" />';
-		echo '<input type="submit" value="'._('Delete access to this shared folder').'" />';
-		echo '</form></td>';
-		echo '</tr>';
-	}
+		echo '<table border="0" cellspacing="1" cellpadding="3">';
+		foreach ($used_sharedfolders as $sharedfolder) {
+			echo '<tr>';
+			echo '<td><a href="sharedfolders.php?action=manage&amp;id='.$sharedfolder->id.'">'.$sharedfolder->name.'</a></td>';
+			echo '<td><form action="actions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to delete this shared folder access?').'\');">';
+			echo '<input type="hidden" name="name" value="SharedFolder_ACL" />';
+			echo '<input type="hidden" name="action" value="del" />';
+			echo '<input type="hidden" name="sharedfolder_id" value="'.$sharedfolder->id.'" />';
+			echo '<input type="hidden" name="usergroup_id" value="'.$group->getUniqueID().'" />';
+			echo '<input type="submit" value="'._('Delete access to this shared folder').'" />';
+			echo '</form></td>';
+			echo '</tr>';
+		}
 
-	if (count($available_sharedfolders) > 0) {
-		echo '<tr><form action="actions.php" method="post"><td>';
-		echo '<input type="hidden" name="name" value="SharedFolder_ACL" />';
-		echo '<input type="hidden" name="action" value="add" />';
-		echo '<input type="hidden" name="usergroup_id" value="'.$group->getUniqueID().'" />';
-		echo '<select name="sharedfolder_id">';
-		foreach($available_sharedfolders as $sharedfolder)
-			echo '<option value="'.$sharedfolder->id.'" >'.$sharedfolder->name.'</option>';
-		echo '</select>';
-		echo '</td><td><input type="submit" value="'._('Add access to this shared folder').'" /></td>';
-		echo '</form></tr>';
-	}
-	echo '</table>';
+		if (count($available_sharedfolders) > 0) {
+			echo '<tr><form action="actions.php" method="post"><td>';
+			echo '<input type="hidden" name="name" value="SharedFolder_ACL" />';
+			echo '<input type="hidden" name="action" value="add" />';
+			echo '<input type="hidden" name="usergroup_id" value="'.$group->getUniqueID().'" />';
+			echo '<select name="sharedfolder_id">';
+			foreach($available_sharedfolders as $sharedfolder)
+				echo '<option value="'.$sharedfolder->id.'" >'.$sharedfolder->name.'</option>';
+			echo '</select>';
+			echo '</td><td><input type="submit" value="'._('Add access to this shared folder').'" /></td>';
+			echo '</form></tr>';
+		}
+		echo '</table>';
+		echo '</div>';
+    }
 
   echo '</div>';
   page_footer();
